+++ GreekReference/src/main/AndroidManifest.xml
@@ -38,24 +38,24 @@
                 android:resource="@xml/searchable" />
         </activity>
         <activity
-            android:name=".views.lexicon.LexiconDetailActivity"
+            android:name=".views.detail.lexicon.LexiconDetailActivity"
             android:parentActivityName=".views.MainActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".views.MainActivity" />
         </activity>
         <activity
-            android:name=".views.syntax.SyntaxDetailActivity"
+            android:name=".views.detail.syntax.SyntaxDetailActivity"
             android:parentActivityName=".views.MainActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".views.MainActivity" />
         </activity>
         <activity
-            android:name=".SettingsActivity"
+            android:name=".views.SettingsActivity"
             android:label="@string/title_activity_settings">
         </activity>
-        <activity android:name=".lexicon.PerseusToolActivity">
+        <activity android:name=".views.PerseusToolActivity">
         </activity>
 
         <provider
