+++ GreekReference/src/main/java/com/benlinskey/greekreference/GreekTextView.java
@@ -44,6 +44,9 @@ import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.widget.TextView;
 
+import com.benlinskey.greekreference.R;
+import com.benlinskey.greekreference.TextSize;
+
 /**
  * A custom {@link TextView} that uses a font capable of displaying polytonic Greek characters.
  * <p>
