+++ GreekReference/src/main/java/com/benlinskey/greekreference/presenters/SyntaxPresenter.java
@@ -28,7 +28,7 @@ import android.widget.ListView;
 import com.benlinskey.greekreference.R;
 import com.benlinskey.greekreference.data.appdata.AppDataContract;
 import com.benlinskey.greekreference.data.syntax.SyntaxContract;
-import com.benlinskey.greekreference.views.syntax.SyntaxDetailView;
+import com.benlinskey.greekreference.views.detail.syntax.SyntaxDetailView;
 
 public class SyntaxPresenter {
