+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/MainActivity.java
@@ -27,7 +27,6 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
@@ -36,33 +35,31 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.widget.SearchView;
 import android.support.v7.widget.Toolbar;
-import android.view.KeyEvent;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.Toast;
 
-import com.benlinskey.greekreference.AbstractDetailFragment;
-import com.benlinskey.greekreference.AbstractListFragment;
+import com.benlinskey.greekreference.views.detail.AbstractDetailFragment;
+import com.benlinskey.greekreference.views.list.AbstractListFragment;
 import com.benlinskey.greekreference.Mode;
 import com.benlinskey.greekreference.R;
-import com.benlinskey.greekreference.lexicon.AbstractLexiconListFragment;
-import com.benlinskey.greekreference.lexicon.LexiconBrowseListFragment;
-import com.benlinskey.greekreference.lexicon.LexiconFavoritesListFragment;
-import com.benlinskey.greekreference.lexicon.LexiconHistoryListFragment;
+import com.benlinskey.greekreference.views.list.lexicon.AbstractLexiconListFragment;
+import com.benlinskey.greekreference.views.list.lexicon.LexiconBrowseListFragment;
+import com.benlinskey.greekreference.views.list.lexicon.LexiconFavoritesListFragment;
+import com.benlinskey.greekreference.views.list.lexicon.LexiconHistoryListFragment;
 import com.benlinskey.greekreference.navigationdrawer.NavigationDrawerFragment;
 import com.benlinskey.greekreference.presenters.LexiconPresenter;
 import com.benlinskey.greekreference.presenters.MainPresenter;
 import com.benlinskey.greekreference.presenters.SyntaxPresenter;
-import com.benlinskey.greekreference.syntax.AbstractSyntaxListFragment;
-import com.benlinskey.greekreference.syntax.SyntaxBookmarksListFragment;
-import com.benlinskey.greekreference.syntax.SyntaxBrowseListFragment;
-import com.benlinskey.greekreference.views.lexicon.LexiconDetailActivity;
-import com.benlinskey.greekreference.views.lexicon.LexiconDetailFragment;
-import com.benlinskey.greekreference.views.lexicon.LexiconDetailView;
-import com.benlinskey.greekreference.views.syntax.SyntaxDetailActivity;
-import com.benlinskey.greekreference.views.syntax.SyntaxDetailFragment;
-import com.benlinskey.greekreference.views.syntax.SyntaxDetailView;
+import com.benlinskey.greekreference.views.list.syntax.AbstractSyntaxListFragment;
+import com.benlinskey.greekreference.views.list.syntax.SyntaxBookmarksListFragment;
+import com.benlinskey.greekreference.views.list.syntax.SyntaxBrowseListFragment;
+import com.benlinskey.greekreference.views.detail.lexicon.LexiconDetailActivity;
+import com.benlinskey.greekreference.views.detail.lexicon.LexiconDetailFragment;
+import com.benlinskey.greekreference.views.detail.lexicon.LexiconDetailView;
+import com.benlinskey.greekreference.views.detail.syntax.SyntaxDetailActivity;
+import com.benlinskey.greekreference.views.detail.syntax.SyntaxDetailFragment;
+import com.benlinskey.greekreference.views.detail.syntax.SyntaxDetailView;
 
 /**
  * The app's primary activity. On tablets, this activity displays a two-pane layout containing an
