+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/list/lexicon/LexiconFavoritesListFragment.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.benlinskey.greekreference.lexicon;
+package com.benlinskey.greekreference.views.list.lexicon;
 
 import android.app.Activity;
 import android.app.LoaderManager;
