+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/list/lexicon/LexiconHistoryListFragment.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.benlinskey.greekreference.syntax;
+package com.benlinskey.greekreference.views.list.lexicon;
 
 import android.app.Activity;
 import android.app.LoaderManager;
@@ -22,7 +22,6 @@ import android.content.ContentResolver;
 import android.content.CursorLoader;
 import android.content.Loader;
 import android.database.Cursor;
-import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ListView;
@@ -30,39 +29,29 @@ import android.widget.SimpleCursorAdapter;
 
 import com.benlinskey.greekreference.R;
 import com.benlinskey.greekreference.data.appdata.AppDataContract;
+import com.benlinskey.greekreference.data.appdata.LexiconHistoryProvider;
 
 /**
- * A {@link AbstractSyntaxListFragment} used to display a list of all words stored in the syntax bookmarks
- * list.
+ * A {@link AbstractLexiconListFragment} used to display a list of all words stored in
+ * the lexicon history list.
  */
-public class SyntaxBookmarksListFragment extends AbstractSyntaxListFragment
-        implements LoaderManager.LoaderCallbacks<Cursor> {
+public class LexiconHistoryListFragment extends AbstractLexiconListFragment
+        implements LoaderManager.LoaderCallbacks<Cursor>{
+    public static final String NAME = "lexicon_History";
 
-    public static final String NAME = "syntax_bookmarks";
-
-    private static final String[] PROJECTION = new String[] {
-        AppDataContract.SyntaxBookmarks._ID,
-        AppDataContract.SyntaxBookmarks.COLUMN_NAME_SYNTAX_SECTION,
-        AppDataContract.SyntaxBookmarks.COLUMN_NAME_SYNTAX_ID
-    };
+    private static final String[] PROJECTION = new String[] {AppDataContract.LexiconHistory._ID,
+            AppDataContract.LexiconHistory.COLUMN_NAME_WORD};
     private static final String SELECTION = "";
     private static final String[] SELECTION_ARGS = {};
-    private static final String SORT_ORDER = 
-            AppDataContract.SyntaxBookmarks.COLUMN_NAME_SYNTAX_ID + " ASC";
-
-    /**
-     * The serialization (saved instance state) Bundle key representing the
-     * activated item position. Only used on tablets.
-     */
-    private static final String STATE_ACTIVATED_POSITION = "activated_position";
+    private static final String ORDER_BY = AppDataContract.LexiconHistory._ID + " DESC";
 
     private SimpleCursorAdapter mAdapter;
-    
-    /** The fragment's current callback object, which is notified of list item clicks. */
-    private Callbacks mCallbacks = sDummyCallbacks;
 
-    /** The current activated item position. Only used on tablets. */
-    private int mActivatedPosition = ListView.INVALID_POSITION;
+    /**
+     * The fragment's current callback object, which is notified of list item
+     * clicks.
+     */
+    private Callbacks mCallbacks = sDummyCallbacks;
 
     /**
      * A dummy implementation of the {@link Callbacks} interface that does
@@ -70,54 +59,46 @@ public class SyntaxBookmarksListFragment extends AbstractSyntaxListFragment
      */
     private static final Callbacks sDummyCallbacks = new Callbacks() {
         @Override
-        public void onItemSelected(String fragmentName) {}
+        public void onItemSelected(String fragmentName) {
+        }
     };
 
     /**
      * Mandatory empty constructor for the fragment manager to instantiate the
      * fragment (e.g. upon screen orientation changes).
      */
-    public SyntaxBookmarksListFragment() {}
+    public LexiconHistoryListFragment() {
+    }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        // TODO: Add chapter name before section name in this list?
-        String[] fromColumns = {AppDataContract.SyntaxBookmarks.COLUMN_NAME_SYNTAX_SECTION};
+        // Create and set list adapter.
+        String[] fromColumns = {AppDataContract.LexiconHistory.COLUMN_NAME_WORD};
         int[] toViews = {android.R.id.text1};
-        int layout = android.R.layout.simple_list_item_activated_1;
-        mAdapter = new SimpleCursorAdapter(getActivity(), layout, null, fromColumns, toViews, 0);
+        mAdapter = new android.widget.SimpleCursorAdapter(getActivity(),
+                R.layout.greek_simple_list_item_activated_1, null, fromColumns, toViews, 0);
         setListAdapter(mAdapter);
+
         getLoaderManager().initLoader(0, null, this);
     }
 
     @Override
     public Loader<Cursor> onCreateLoader(int id, Bundle args) {
-        return new CursorLoader(getActivity(), AppDataContract.SyntaxBookmarks.CONTENT_URI,
-                PROJECTION, SELECTION, SELECTION_ARGS, SORT_ORDER);
+        return new CursorLoader(getActivity(), LexiconHistoryProvider.CONTENT_URI, PROJECTION, SELECTION,
+                SELECTION_ARGS, ORDER_BY);
     }
 
     @Override
     public void onLoadFinished(Loader<Cursor> loader, Cursor data) {
         mAdapter.swapCursor(data);
-        setNoItemsView(R.string.syntax_bookmarks_empty_view);
+        setNoItemsView(R.string.lexicon_history_empty_view);
     }
 
     @Override
     public void onLoaderReset(Loader<Cursor> loader) {
         mAdapter.swapCursor(null);
-    }
-
-    @Override
-    public void onViewCreated(View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-
-        // Restore the previously serialized activated item position.
-        if (savedInstanceState != null
-                && savedInstanceState.containsKey(STATE_ACTIVATED_POSITION)) {
-            setActivatedPosition(savedInstanceState.getInt(STATE_ACTIVATED_POSITION));
-        }
     }
 
     @Override
@@ -143,36 +124,32 @@ public class SyntaxBookmarksListFragment extends AbstractSyntaxListFragment
     @Override
     public void onListItemClick(ListView listView, View view, int position, long id) {
         super.onListItemClick(listView, view, position, id);
+
         Cursor cursor = (Cursor) mAdapter.getItem(position);
-        int syntaxBookmarksId = cursor.getInt(0);
-        setSelectedSyntaxItemId(syntaxBookmarksId);
-        mCallbacks.onItemSelected(NAME);
+        int lexiconHistoryId = cursor.getInt(0);
+
+        setSelectedLexiconItemId(lexiconHistoryId);
+
+        // Notify the active callbacks interface (the activity, if the
+        // fragment is attached to one) that an item has been selected.
+        mCallbacks.onItemSelected(NAME); // Database IDs start at 1.
     }
 
     @Override
-    public void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-        if (mActivatedPosition != ListView.INVALID_POSITION) {
-            // Serialize and persist the activated item position.
-            outState.putInt(STATE_ACTIVATED_POSITION, mActivatedPosition);
-        }
-    }
-
-    @Override
-    protected void setSelectedSyntaxItemId(int id) {
-        String[] columns = new String[] {AppDataContract.SyntaxBookmarks.COLUMN_NAME_SYNTAX_ID};
-        String selection = AppDataContract.SyntaxBookmarks._ID + " = ?";
+    protected void setSelectedLexiconItemId(int id) {
+        String[] columns = new String[] {AppDataContract.LexiconHistory.COLUMN_NAME_LEXICON_ID};
+        String selection = AppDataContract.LexiconHistory._ID + " = ?";
         String[] selectionArgs = new String[] {Integer.toString(id)};
         ContentResolver resolver = getActivity().getContentResolver();
-        Uri uri = AppDataContract.SyntaxBookmarks.CONTENT_URI;
-        Cursor cursor = resolver.query(uri, columns, selection, selectionArgs, null);
+        Cursor cursor = resolver.query(AppDataContract.LexiconHistory.CONTENT_URI, columns,
+                                       selection, selectionArgs, null);
 
         if (cursor == null) {
             throw new NullPointerException("ContentResolver#query() returned null");
         }
 
         if (cursor.moveToFirst()) {
-            mSelectedSyntaxId = cursor.getInt(0);
+            mSelectedLexiconId = cursor.getInt(0);
             cursor.close();
         } else {
             cursor.close();
