+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/list/syntax/SyntaxBrowseListFragment.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.benlinskey.greekreference.lexicon;
+package com.benlinskey.greekreference.views.list.syntax;
 
 import android.app.Activity;
 import android.app.LoaderManager;
@@ -26,33 +26,30 @@ import android.view.View;
 import android.widget.ListView;
 import android.widget.SimpleCursorAdapter;
 
-import com.benlinskey.greekreference.R;
-import com.benlinskey.greekreference.data.lexicon.LexiconContract;
+import com.benlinskey.greekreference.data.syntax.SyntaxContract;
 
 /**
- * A {@link AbstractLexiconListFragment} used to display a list of all words in the
- * lexicon.
+ * A {@link AbstractSyntaxListFragment} used to display a list of all sections in the Overview of Greek
+ * Syntax text.
  */
-public class LexiconBrowseListFragment extends AbstractLexiconListFragment
+public class SyntaxBrowseListFragment extends AbstractSyntaxListFragment
         implements LoaderManager.LoaderCallbacks<Cursor> {
 
-    public static final String NAME = "lexicon_browse";
-
-    private static final String TAG = "LexiconBrowseListFragment";
-    private static final String[] PROJECTION
-            = new String[] {LexiconContract._ID, LexiconContract.COLUMN_GREEK_FULL_WORD,
-                    LexiconContract.COLUMN_GREEK_LOWERCASE};
-    private static final String SELECTION = "";
-    private static final String[] SELECTION_ARGS = {};
-    private static final String ORDER_BY = LexiconContract.COLUMN_GREEK_LOWERCASE + " ASC";
+    public static final String NAME = "syntax_browse";
 
     private SimpleCursorAdapter mAdapter;
+    private static final String[] PROJECTION = new String[] {
+        SyntaxContract._ID,
+        SyntaxContract.COLUMN_NAME_SECTION
+    };
+    private static final String SELECTION = "";
+    private static final String[] SELECTION_ARGS = {};
 
     /**
-     * The fragment's current callback object, which is notified of list item
-     * clicks.
+     * The serialization (saved instance state) Bundle key representing the
+     * activated item position. Only used on tablets.
      */
-    private Callbacks mCallbacks = sDummyCallbacks;
+    private static final String STATE_ACTIVATED_POSITION = "activated_position";
 
     /**
      * A dummy implementation of the {@link Callbacks} interface that does
@@ -60,35 +57,39 @@ public class LexiconBrowseListFragment extends AbstractLexiconListFragment
      */
     private static final Callbacks sDummyCallbacks = new Callbacks() {
         @Override
-        public void onItemSelected(String fragmentName) {
-        }
+        public void onItemSelected(String fragmentName) {}
     };
+    
+    /** The fragment's current callback object, which is notified of list item clicks. */
+    private Callbacks mCallbacks = sDummyCallbacks;
 
+    /** The current activated item position. Only used on tablets. */
+    private int mActivatedPosition = ListView.INVALID_POSITION;
+    
     /**
      * Mandatory empty constructor for the fragment manager to instantiate the
      * fragment (e.g. upon screen orientation changes).
      */
-    public LexiconBrowseListFragment() {
+    public SyntaxBrowseListFragment() {
     }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        // Create and set list adapter.
-        // TODO: Replace this with a more efficient adapter.
-        String[] fromColumns = {LexiconContract.COLUMN_GREEK_FULL_WORD};
+        // TODO: Create a custom adapter with support for chapter headings.
+        String[] fromColumns = {SyntaxContract.COLUMN_NAME_SECTION};
         int[] toViews = {android.R.id.text1};
-        mAdapter = new SimpleCursorAdapter(getActivity(),
-                R.layout.greek_simple_list_item_activated_1, null, fromColumns, toViews, 0);
+        int layout = android.R.layout.simple_list_item_activated_1;
+        mAdapter = new SimpleCursorAdapter(getActivity(), layout, null, fromColumns, toViews, 0);
         setListAdapter(mAdapter);
         getLoaderManager().initLoader(0, null, this);
     }
 
     @Override
     public Loader<Cursor> onCreateLoader(int id, Bundle args) {
-        return new CursorLoader(getActivity(), LexiconContract.CONTENT_URI, PROJECTION, SELECTION,
-                SELECTION_ARGS, ORDER_BY);
+        return new CursorLoader(getActivity(), SyntaxContract.CONTENT_URI, PROJECTION, SELECTION,
+                SELECTION_ARGS, null);
     }
 
     @Override
@@ -105,8 +106,11 @@ public class LexiconBrowseListFragment extends AbstractLexiconListFragment
     public void onViewCreated(View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        getListView().setFastScrollEnabled(true);
-        getListView().setFastScrollAlwaysVisible(true);
+        // Restore the previously serialized activated item position.
+        if (savedInstanceState != null
+                && savedInstanceState.containsKey(STATE_ACTIVATED_POSITION)) {
+            setActivatedPosition(savedInstanceState.getInt(STATE_ACTIVATED_POSITION));
+        }
     }
 
     @Override
@@ -132,27 +136,21 @@ public class LexiconBrowseListFragment extends AbstractLexiconListFragment
     @Override
     public void onListItemClick(ListView listView, View view, int position, long id) {
         super.onListItemClick(listView, view, position, id);
-
-        Cursor cursor = (Cursor) mAdapter.getItem(position);
-        int lexiconId = cursor.getInt(0);
-        setSelectedLexiconItemId(lexiconId);
-        mCallbacks.onItemSelected(NAME); // Positions are off by one from database ID.
+        setSelectedSyntaxItemId(position);
+        mCallbacks.onItemSelected(NAME);
     }
 
     @Override
-    protected void setSelectedLexiconItemId(int id) {
-        mSelectedLexiconId = id;
+    public void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        if (mActivatedPosition != ListView.INVALID_POSITION) {
+            // Serialize and persist the activated item position.
+            outState.putInt(STATE_ACTIVATED_POSITION, mActivatedPosition);
+        }
     }
 
-    /**
-     * Selects the specified item.
-     * @param id the lexicon database ID of the item to select
-     */
-    public void selectItem(int id) {
-        mSelectedLexiconId = id;
-        int position = id - 1;
-        setActivatedPosition(position);
-        getListView().setSelection(position);
-        mCallbacks.onItemSelected(NAME);
+    @Override
+    protected void setSelectedSyntaxItemId(int id) {
+        mSelectedSyntaxId = id + 1;
     }
 }
